perf: extract closing host rendering in separate component
